backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
langchain-anthropic==0.3.4
langchain-google-genai==2.0.8
chromadb==0.5.23
openai==1.59.3
anthropic==0.42.0
google-generativeai==0.8.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.32.3
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
sentence-transformers==3.3.1
pypdf==5.1.0
tiktoken==0.8.0
supabase==2.11.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
langsmith==0.2.10
